requirements.txt
ipywidgets==8.1.2
matplotlib==3.8.4
pandas==2.2.2
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
Package                   Version
------------------------- -----------
agate                     1.9.1
annotated-types           0.7.0
attrs                     23.2.0
Babel                     2.15.0
cachetools                5.4.0
certifi                   2024.7.4
charset-normalizer        3.3.2
click                     8.1.7
colorama                  0.4.6
daff                      1.3.46
db-dtypes                 1.2.0
dbt-adapters              1.3.2
dbt-bigquery              1.8.2
dbt-common                1.6.0
dbt-core                  1.8.4
dbt-extractor             0.5.1
dbt-semantic-interfaces   0.5.1
deepdiff                  7.0.1
google-api-core           2.19.1
google-auth               2.32.0
google-cloud-bigquery     3.25.0
google-cloud-core         2.4.1
google-cloud-dataproc     5.10.1
google-cloud-storage      2.18.0
google-crc32c             1.5.0
google-resumable-media    2.7.1
googleapis-common-protos  1.63.2
grpc-google-iam-v1        0.13.1
grpcio                    1.65.1
grpcio-status             1.62.2
idna                      3.7
importlib-metadata        6.11.0
isodate                   0.6.1
Jinja2                    3.1.4
jsonschema                4.23.0
jsonschema-specifications 2023.12.1
leather                   0.4.0
Logbook                   1.5.3
MarkupSafe                2.1.5
mashumaro                 3.13.1
minimal-snowplow-tracker  0.0.2
more-itertools            10.3.0
msgpack                   1.0.8
networkx                  3.3
numpy                     2.0.1
ordered-set               4.1.0
packaging                 24.1
pandas                    2.2.2
parsedatetime             2.6
pathspec                  0.12.1
pip                       23.0.1
proto-plus                1.24.0
protobuf                  4.25.4
pyarrow                   17.0.0
pyasn1                    0.6.0
pyasn1_modules            0.4.0
pydantic                  2.8.2
pydantic_core             2.20.1
python-dateutil           2.9.0.post0
python-slugify            8.0.4
pytimeparse               1.1.8
pytz                      2024.1
PyYAML                    6.0.1
referencing               0.35.1
requests                  2.32.3
rpds-py                   0.19.1
rsa                       4.9
setuptools                65.5.0
six                       1.16.0
sqlparse                  0.5.1
text-unidecode            1.3
typing_extensions         4.12.2
tzdata                    2024.1
urllib3                   2.2.2
zipp                      3.19.2
